requirements.txt
Django==5.2.1
django-crispy-forms
django-crispy-bootstrap4
django-widget-tweaks
requests>=2.25.0
Pillow